File working, but error with insert or delete, maybe. The important change was in graph.cpp I added an extra node to subsets

+++ inc/graph.h
@@ -17,7 +17,7 @@ enum color {WHITE, GRAY, BLACK}; //simple enum for the color tracking in the bfs
  */ 
 typedef struct edge {
     int src, dest;
-    float weight;
+    int weight;
 
     bool operator==(const edge& e) {
         return ((src == e.src) && (dest == e.dest) && (weight == e.weight))
